requirements.txt
aniso8601==9.0.1
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
Flask==2.3.2
graphene==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.3.3
python-decouple==3.8
python-dotenv==1.0.0
Werkzeug==2.3.6
